Updates, running into a problem.
Added statements to capture contestant names to str[] names, and to add
contestant talents to str[] talents.
Created method makeArray() to display names and talents as a table.
Commented out the program's function that requests the user enter a
talent code to to see the number of contestans using that talent.
Invoked makeArray(names, talents); in its place.